Update label for Docker Hub password parameter
Updating the parameter label for "ReferenceArchitecture.Eks.Docker.Password" to be "Docker Hub Access Token." This parameter is used to create a secret in the GitHub repo which is used by the workflow file that is generated. The login for the Docker GitHub actions should use an access token and not the account login password per: https://github.com/docker/login-action#docker-hub.